Add explicit support for AArch64
Specifies register shape and suffix for floating point and vector
registers. The former cannot be benchmarked without and the latter would
require adding the required suffixes manually in the instruction
operands. Doing both allows using them in the same manner as on x86.

Additionally there are two small changes affecting all architectures:

Allow the 'w' constraint code, which is used for vector registers on
aarch64.

Always specify a clobber for the flags register as many instructions one
might want to benchmark modify it.